+++ src/component/mypage/FileUpload.tsx
@@ -4,12 +4,13 @@ import { URL } from 'config';
 import styled from 'styled-components';
 
 interface Props {
-  imgStore: (img: String) => void;
+  image: string;
 }
 
 const FileUpload: React.FC<Props> = (props: Props) => {
   const [img, setImg] = useState<any>();
   const [imgUpload, setImgUpload] = useState<any>();
+  // console.log(props.image);
 
   const isSelectedImg = (event: React.ChangeEvent<HTMLInputElement>) => {
     console.log('files', event.target.files);
@@ -23,36 +24,44 @@ const FileUpload: React.FC<Props> = (props: Props) => {
   ) => {
     const fd = new FormData();
     fd.append('filename', img);
-    console.log(fd);
+    // console.log('formdata', fd);
 
     const token = window.sessionStorage.getItem('token');
-    console.log(token);
+    // console.log(token);
 
     axios
       .post(`${URL}/user/profile-upload`, fd, {
-        headers: { token: token },
+        headers: { Authorization: token },
       })
-      //백엔드 api주소 넣기
       .then(res => {
-        // console.log(res);
-        console.log(res.data.image_url);
+        console.log(res);
+        console.log('이미지 주소', res.data.image_url);
         setImgUpload(res.data.image_url);
-        props.imgStore(res.data.image_url);
+
+        if (token) {
+          fetch(`${URL}/user/image`, {
+            method: 'POST',
+            headers: { Authorization: token },
+            body: JSON.stringify({
+              image_url: res.data.image_url,
+            }),
+          }).then(res => console.log(res));
+        }
       })
+
       .catch(error => {
         console.log(error.response);
       });
   };
   return (
     <Wrapper>
-      <ImgBox>
+      <ImgBox style={{ backgroundImage: `url(${props.image})` }}>
         <Img src={imgUpload} alt="" />
-        {/* http://localhost:3001/images/gopizza.png */}
       </ImgBox>
       <InputBox>
         <ImgInput
           type="file"
-          accept="image/png, image/jpeg, image/jpg"
+          accept=" image/jpeg, image/jpg"
           onChange={isSelectedImg}
         />
         <ImgBtn type="submit" onClick={isUploadedImg}>
@@ -76,7 +85,8 @@ const ImgBox = styled.div`
   width: 100px;
   height: 100px;
   border-radius: 50%;
-  background-image: url('https://www.nelson-chambers.co.uk/front/images/default-user.jpg');
+  /* background-image: url('https://www.nelson-chambers.co.uk/front/images/default-user.jpg'); */
+  background-color: lightgray;
   background-size: 100%;
 `;
 const Img = styled.img`
